use `test-utilities` instead of `@itwin/presentation-testing` in `learning-snippets`
