+++ apps/learning-snippets/package.json
@@ -39,7 +39,6 @@
     "@itwin/presentation-frontend": "^5.7.0",
     "@itwin/presentation-hierarchies": "^1.7.9",
     "@itwin/presentation-shared": "^1.2.8",
-    "@itwin/presentation-testing": "^5.4.14",
     "@itwin/property-grid-react": "workspace:*",
     "@itwin/quantity-formatting-react": "workspace:*",
     "@itwin/tree-widget-react": "workspace:*",
@@ -56,12 +55,12 @@
     "eslint-config-prettier": "^10.1.8",
     "eslint-plugin-react": "^7.37.5",
     "eslint-plugin-unused-imports": "^4.3.0",
-    "fast-xml-parser": "^5.7.0",
     "happy-dom": "^20.8.9",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "rimraf": "^6.1.2",
     "sass-embedded": "1.98.0",
+    "test-utilities": "workspace:*",
     "typescript": "~5.9.3",
     "vitest": "^4.1.2"
   },
