+++ apps/learning-snippets/src/utils/InitializationUtils.ts
@@ -3,20 +3,18 @@
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
 
-import * as fs from "fs";
 import * as path from "path";
 import * as url from "url";
 import { IModelReadRpcInterface } from "@itwin/core-common";
 import { ECSchemaRpcInterface } from "@itwin/ecschema-rpcinterface-common";
 import { ECSchemaRpcImpl } from "@itwin/ecschema-rpcinterface-impl";
 import { PresentationRpcInterface } from "@itwin/presentation-common";
-import { HierarchyCacheMode, initialize as initializePresentationTesting, terminate as terminatePresentationTesting } from "@itwin/presentation-testing";
+import { initializeCore, terminateCore } from "test-utilities";
+import { HierarchyCacheMode } from "@itwin/presentation-backend";
 
 export async function initializeLearningSnippetsTests() {
-  const outDir = path.join(path.dirname(url.fileURLToPath(import.meta.url)), "..", "..", "lib", "test", "out", `${process.pid}`);
-  fs.mkdirSync(outDir, { recursive: true });
-
-  await initializePresentationTesting({
+  await initializeCore({
+    testOutputDir: path.join(path.dirname(url.fileURLToPath(import.meta.url)), "..", "..", "lib", "test", "out", `${process.pid}`),
     backendProps: {
       caching: {
         hierarchies: {
@@ -25,10 +23,6 @@ export async function initializeLearningSnippetsTests() {
         },
       },
     },
-    testOutputDir: path.join(outDir, "output"),
-    backendHostProps: {
-      cacheDir: path.join(outDir, "cache"),
-    },
     rpcs: [IModelReadRpcInterface, PresentationRpcInterface, ECSchemaRpcInterface],
   });
   // eslint-disable-next-line @itwin/no-internal
@@ -36,5 +30,5 @@ export async function initializeLearningSnippetsTests() {
 }
 
 export async function terminateLearningSnippetsTests() {
-  await terminatePresentationTesting();
+  await terminateCore();
 }
