+++ packages/itwin/test-utilities/package.json
@@ -26,7 +26,9 @@
   },
   "dependencies": {
     "@itwin/core-backend": "^5.7.0",
-    "@itwin/presentation-backend": "^5.7.0"
+    "@itwin/presentation-backend": "^5.7.0",
+    "fast-xml-parser": "^5.7.0",
+    "sanitize-filename": "^1.6.4"
   },
   "devDependencies": {
     "@itwin/appui-abstract": "^5.7.0",
